swarm-art/swarm-gen-game-art-main-main/requirements.txt
Pillow>=10.0.0
numpy>=1.20.0
PyQt6>=6.5.0
scipy>=1.10.0
